requirements.txt
failed to get console mode for stdout: The handle is invalid.
asgiref==3.8.1
Babel==2.15.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.2
django-allauth==0.61.1
django-cache-url==3.4.5
django-crispy-forms==2.1
django-jalali-date==1.1.2
django-rosetta==0.10.0
environs==11.0.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
marshmallow==3.21.1
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
polib==1.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
